This changelist is a set of breaking changes for earlier Lightyear users.
This changes refactors some of the functions in Lightyear.Strings to
a new module Lightyear.Char, to follow more closely the implementation
of Daan Leijen's Parsec. Additionally it adds missing functionality from
that library, found in Parsec.Text.Char.

All tests have been updated to account for these changes and run as
expected.

The breaking changes are small and require adding an additional
import:

    import Lightyear.Char